+++ moneysocial/app/Routes/MainNavigation.js
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
 import SignInScreen from '../screens/WelcomeScreen';
 import WelcomeScreen from '../screens/WelcomeScreen';
-import PersonalExpense from '../screens/PersonalExpense';
-import groupExpense from '../screens/groupExpense';
+import HomeScreen from '../screens/HomeScreen';
+import groupList from '../screens/groupList';
 import ExpenseListScreen from '../screens/ExpenseListScreen';
 import { NavigationContainer } from '@react-navigation/native';
 
@@ -15,9 +15,9 @@ function stackNavigator() {
         <NavigationContainer>
             <Stack.Navigator>
                         <Stack.Screen name ="WelcomeScreen" component={WelcomeScreen}/>
-                        <Stack.Screen name ="PersonalExpense" component={PersonalExpense}/>
+                        <Stack.Screen name ="HomeScreen" component={HomeScreen}/>
                         
-                        <Stack.Screen name ="groupExpense" component={groupExpense}/>
+                        <Stack.Screen name ="groupList" component={groupList}/>
                         <Stack.Screen name ="ExpenseListScreen" component={ExpenseListScreen}/>
             </Stack.Navigator>
         </NavigationContainer>
